Split cli command code
