+++ pipelex/cli/commands/__init__.py
@@ -1,0 +1,11 @@
+"""Command groups for Pipelex CLI.
+
+This package organizes CLI commands into logical modules.
+"""
+
+from .init_cmd import init_app
+from .migrate_cmd import migrate_app
+from .show_cmd import show_app
+from .validate_cmd import validate_app
+
+__all__ = ["init_app", "migrate_app", "show_app", "validate_app"]
